requirements.txt
secure-smtplib
fastapi
colorama
python-jose
python-dotenv
sendgrid